refactor(layout): extract the inline token model out of TextFlowSupport
Move the inline layout tokens — the sealed InlineLayoutToken and its
InlineText / Image / Shape / Svg record variants (with the SVG-layer lowering)
— out of the TextFlowSupport god-class into their own package-private file.
The types were private nested; as top-level package-private types the wrapping
loops reference them by simple name unchanged. Pure type move, no behaviour
change; the paragraph wrapping loops that build and consume them follow next.
